Support for ALPINE_NO_RESOLVER env var
When this env var is passed to the container, then the container DNS
server will not be started, but resolv.conf will still be changed to
have the nameserver 127.0.0.1. This is a specific workaround making it
possible to run multiple Alpine-Kubernetes images in a single Pod that
shares a common network port stack so that only one DNS server can bind
to localhost port 53.